refactor out panels
